Vastly improved door interaction
Added Lerp to limit vibration of the door, as well as overall improvements to the interaction of a door. I think I found the source of the issue causing the doors to misbehave. However, the issue still remains of opening a door from the inside (door swinging out). It is difficult for the player to open the door in these cases.